Move connection to the movementservice to BotActivity
