+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/human/HumanActivity.java
@@ -1,12 +1,7 @@
 package com.novoda.tpbot.human;
 
-import android.content.ComponentName;
-import android.content.Context;
-import android.content.Intent;
-import android.content.ServiceConnection;
 import android.os.Bundle;
 import android.os.Handler;
-import android.os.IBinder;
 import android.support.v7.app.AppCompatActivity;
 
 import com.novoda.notils.caster.Views;
@@ -14,7 +9,6 @@ import com.novoda.tpbot.Direction;
 import com.novoda.tpbot.R;
 import com.novoda.tpbot.support.SelfDestructingMessageView;
 import com.novoda.tpbot.support.SwitchableView;
-import com.novoda.tpbot.bot.MovementService;
 
 import java.util.concurrent.TimeUnit;
 
@@ -25,8 +19,6 @@ public class HumanActivity extends AppCompatActivity {
 
     private SelfDestructingMessageView debugView;
     private SwitchableView switchableView;
-    private MovementService movementService;
-    private boolean bound;
 
     private Handler handler;
     private String currentCommand;
@@ -50,15 +42,15 @@ public class HumanActivity extends AppCompatActivity {
 
     private final ControllerListener controllerListener = new ControllerListener() {
 
-            @Override
-            public void onDirectionPressed(Direction direction) {
-                startRepeatingCommand(direction.rawCommand());
-            }
+        @Override
+        public void onDirectionPressed(Direction direction) {
+            startRepeatingCommand(direction.rawCommand());
+        }
 
-            @Override
-            public void onDirectionReleased(Direction direction) {
-                stopRepeatingCommand(direction.rawCommand());
-            }
+        @Override
+        public void onDirectionReleased(Direction direction) {
+            stopRepeatingCommand(direction.rawCommand());
+        }
 
         @Override
         public void onLazersFired() {
@@ -94,16 +86,7 @@ public class HumanActivity extends AppCompatActivity {
 
     private void sendCommand(String command) {
         debugView.showTimed(command, COMMAND_REPEAT_DELAY);
-        if (bound) {
-            movementService.sendCommand(command);
-        }
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-        Intent intent = new Intent(this, MovementService.class);
-        bindService(intent, serviceConnection, Context.BIND_AUTO_CREATE);
+        // TODO: send command to the receiver (bot) part
     }
 
     @Override
@@ -111,31 +94,6 @@ public class HumanActivity extends AppCompatActivity {
         stopRepeatingCommand(currentCommand);
         super.onPause();
     }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-        if (bound) {
-            unbindService(serviceConnection);
-            bound = false;
-        }
-    }
-
-    private ServiceConnection serviceConnection = new ServiceConnection() {
-
-        @Override
-        public void onServiceConnected(ComponentName className,
-                                       IBinder iBinder) {
-            MovementService.Binder binder = (MovementService.Binder) iBinder;
-            movementService = binder.getService();
-            bound = true;
-        }
-
-        @Override
-        public void onServiceDisconnected(ComponentName arg0) {
-            bound = false;
-        }
-    };
 
     private final ServerDeclarationListener serverDeclarationListener = new ServerDeclarationListener() {
         @Override
